+++ CHANGELOG.md
@@ -57,6 +57,7 @@
 - Refactor food basked creation to vue and rest api !2906 @AntonBallmaier
 - Replaced the maps on basket pages with vue and removed vMap !2880 @alex.simm
 - Modernized RegionControl to a Symfony controller !2741 !2945 @\_fridtjof_
+- Modernized StatisticsControl to a Symfony controller !2953 @\_fridtjof_
 
 ## Dev/Test/CI stuff
 - removed CHANGELOG.md merge=union !2866 @chriswalg